GEODE-5945: Updated CommandManagerJUnitTest to accurately test loading commands
The testCommandManagerLoadCommands test in CommandManagerJUnitTest was
only confirming that the number of commands loaded was greater than 0.
This failed to raise issues if there were problems loading commands found in
command markers. We have modified the test to use the ClassPathLoader
and ClassPathScanLoadHelper to validate that the number of commands
loaded is correct.

[GEODE-5945]